backup_pre_refactor_20251203_023425/requirements.txt
flask
flask_sqlalchemy
Flask-Login
Flask-Bcrypt
pg8000
pandas
openpyxl
reportlab
flask-wtf
wtforms
wtforms-sqlalchemy
waitress
requests
python-dotenv
python-docx
PyPDF2
Flask-Moment
Flask-Mail
Flask-Migrate
Email-validator
psycopg2-binary
pg8000